+++ migrations/Version20200205115244.php
@@ -29,45 +29,53 @@ final class Version20200205115244 extends AbstractMigration
 
     public function up(Schema $schema): void
     {
-        $migrates = [
-            ['kimai2_activities', 'activity_id', 'kimai2_activities_rates'],
-            ['kimai2_projects', 'project_id', 'kimai2_projects_rates'],
-            ['kimai2_customers', 'customer_id', 'kimai2_customers_rates'],
-        ];
-
-        foreach ($migrates as $migrateOpts) {
-            $tableName = $migrateOpts[0];
-            $fieldName = $migrateOpts[1];
-            $targetTable = $migrateOpts[2];
-
-            $rules = $this->connection->prepare(
-                'SELECT id, fixed_rate, hourly_rate FROM ' . $tableName . ' WHERE fixed_rate IS NOT NULL OR hourly_rate IS NOT NULL'
-            );
-            $result = $rules->executeQuery();
-
-            foreach ($result->fetchAllAssociative() as $rateRule) {
-                $isFixed = $rateRule['fixed_rate'] !== null;
-                $rate = $rateRule['fixed_rate'] ?? $rateRule['hourly_rate'];
-                $params = ['user_id' => null, $fieldName => $rateRule['id'], 'rate' => $rate, 'fixed' => $isFixed];
-
-                $this->connection->insert($targetTable, $params, ['fixed' => ParameterType::BOOLEAN]);
+        if ($this->isPlatformMySQL()) {
+            $migrates = [
+                ['kimai2_activities', 'activity_id', 'kimai2_activities_rates'],
+                ['kimai2_projects', 'project_id', 'kimai2_projects_rates'],
+                ['kimai2_customers', 'customer_id', 'kimai2_customers_rates'],
+            ];
+    
+            foreach ($migrates as $migrateOpts) {
+                $tableName = $migrateOpts[0];
+                $fieldName = $migrateOpts[1];
+                $targetTable = $migrateOpts[2];
+    
+                $rules = $this->connection->prepare(
+                    'SELECT id, fixed_rate, hourly_rate FROM ' . $tableName . ' WHERE fixed_rate IS NOT NULL OR hourly_rate IS NOT NULL'
+                );
+                $result = $rules->executeQuery();
+    
+                foreach ($result->fetchAllAssociative() as $rateRule) {
+                    $isFixed = $rateRule['fixed_rate'] !== null;
+                    $rate = $rateRule['fixed_rate'] ?? $rateRule['hourly_rate'];
+                    $params = ['user_id' => null, $fieldName => $rateRule['id'], 'rate' => $rate, 'fixed' => $isFixed];
+    
+                    $this->connection->insert($targetTable, $params, ['fixed' => ParameterType::BOOLEAN]);
+                }
             }
+    
+            $schema->getTable('kimai2_customers')->dropColumn('fixed_rate')->dropColumn('hourly_rate');
+            $schema->getTable('kimai2_projects')->dropColumn('fixed_rate')->dropColumn('hourly_rate');
+            $schema->getTable('kimai2_activities')->dropColumn('fixed_rate')->dropColumn('hourly_rate');
+        } else {
+            $this->preventEmptyMigrationWarning();
         }
-
-        $schema->getTable('kimai2_customers')->dropColumn('fixed_rate')->dropColumn('hourly_rate');
-        $schema->getTable('kimai2_projects')->dropColumn('fixed_rate')->dropColumn('hourly_rate');
-        $schema->getTable('kimai2_activities')->dropColumn('fixed_rate')->dropColumn('hourly_rate');
     }
 
     public function down(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE kimai2_customers ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
-        $this->addSql('ALTER TABLE kimai2_customers ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
+        if ($this->isPlatformMySQL()) {
+            $this->addSql('ALTER TABLE kimai2_customers ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
+            $this->addSql('ALTER TABLE kimai2_customers ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
 
-        $this->addSql('ALTER TABLE kimai2_projects ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
-        $this->addSql('ALTER TABLE kimai2_projects ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
+            $this->addSql('ALTER TABLE kimai2_projects ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
+            $this->addSql('ALTER TABLE kimai2_projects ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
 
-        $this->addSql('ALTER TABLE kimai2_activities ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
-        $this->addSql('ALTER TABLE kimai2_activities ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
+            $this->addSql('ALTER TABLE kimai2_activities ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
+            $this->addSql('ALTER TABLE kimai2_activities ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
+        } else {
+            $this->preventEmptyMigrationWarning();
+        }
     }
 }
