+++ migrations/Version20210316224358.php
@@ -28,19 +28,27 @@ final class Version20210316224358 extends AbstractMigration
 
     public function up(Schema $schema): void
     {
-        $bookmarks = $schema->createTable('kimai2_bookmarks');
-        $bookmarks->addColumn('id', 'integer', ['length' => 11, 'autoincrement' => true, 'notnull' => true]);
-        $bookmarks->addColumn('user_id', 'integer', ['length' => 11, 'notnull' => true]);
-        $bookmarks->addColumn('type', 'string', ['length' => 20, 'notnull' => true]);
-        $bookmarks->addColumn('name', 'string', ['length' => 50, 'notnull' => true]);
-        $bookmarks->addColumn('content', 'text', ['notnull' => true]);
-        $bookmarks->addForeignKeyConstraint('kimai2_users', ['user_id'], ['id'], ['onDelete' => 'CASCADE'], 'FK_4016EF25A76ED395');
-        $bookmarks->addUniqueIndex(['user_id', 'name'], 'UNIQ_4016EF25A76ED3955E237E06');
-        $bookmarks->setPrimaryKey(['id']);
+        if ($this->isPlatformMySQL()) {
+            $bookmarks = $schema->createTable('kimai2_bookmarks');
+            $bookmarks->addColumn('id', 'integer', ['length' => 11, 'autoincrement' => true, 'notnull' => true]);
+            $bookmarks->addColumn('user_id', 'integer', ['length' => 11, 'notnull' => true]);
+            $bookmarks->addColumn('type', 'string', ['length' => 20, 'notnull' => true]);
+            $bookmarks->addColumn('name', 'string', ['length' => 50, 'notnull' => true]);
+            $bookmarks->addColumn('content', 'text', ['notnull' => true]);
+            $bookmarks->addForeignKeyConstraint('kimai2_users', ['user_id'], ['id'], ['onDelete' => 'CASCADE'], 'FK_4016EF25A76ED395');
+            $bookmarks->addUniqueIndex(['user_id', 'name'], 'UNIQ_4016EF25A76ED3955E237E06');
+            $bookmarks->setPrimaryKey(['id']);
+        } else {
+            $this->preventEmptyMigrationWarning();
+        }
     }
 
     public function down(Schema $schema): void
     {
-        $this->addSql('DROP TABLE kimai2_bookmarks');
+        if ($this->isPlatformMySQL()) {
+            $this->addSql('DROP TABLE kimai2_bookmarks');
+        } else {
+            $this->preventEmptyMigrationWarning();
+        }
     }
 }
