+++ migrations/Version20210802174318.php
@@ -26,21 +26,29 @@ final class Version20210802174318 extends AbstractMigration
 
     public function up(Schema $schema): void
     {
-        $teamMember = $schema->getTable('kimai2_users_teams');
-        $teamMember->addColumn('id', 'integer', ['autoincrement' => true, 'notnull' => true]);
-        $teamMember->addColumn('teamlead', 'boolean', ['notnull' => true, 'default' => false]);
-        $teamMember->dropPrimaryKey();
-        $teamMember->setPrimaryKey(['id']);
-        $teamMember->addUniqueIndex(['user_id', 'team_id'], 'UNIQ_B5E92CF8A76ED395296CD8AE');
+        if ($this->isPlatformMySQL()) {
+            $teamMember = $schema->getTable('kimai2_users_teams');
+            $teamMember->addColumn('id', 'integer', ['autoincrement' => true, 'notnull' => true]);
+            $teamMember->addColumn('teamlead', 'boolean', ['notnull' => true, 'default' => false]);
+            $teamMember->dropPrimaryKey();
+            $teamMember->setPrimaryKey(['id']);
+            $teamMember->addUniqueIndex(['user_id', 'team_id'], 'UNIQ_B5E92CF8A76ED395296CD8AE');
+        } else {
+            $this->preventEmptyMigrationWarning();
+        }
     }
 
     public function down(Schema $schema): void
     {
-        $teamMember = $schema->getTable('kimai2_users_teams');
-        $teamMember->dropIndex('UNIQ_B5E92CF8A76ED395296CD8AE');
-        $teamMember->dropPrimaryKey();
-        $teamMember->dropColumn('teamlead');
-        $teamMember->dropColumn('id');
-        $teamMember->setPrimaryKey(['user_id', 'team_id']);
+        if ($this->isPlatformMySQL()) {
+            $teamMember = $schema->getTable('kimai2_users_teams');
+            $teamMember->dropIndex('UNIQ_B5E92CF8A76ED395296CD8AE');
+            $teamMember->dropPrimaryKey();
+            $teamMember->dropColumn('teamlead');
+            $teamMember->dropColumn('id');
+            $teamMember->setPrimaryKey(['user_id', 'team_id']);
+        } else {
+            $this->preventEmptyMigrationWarning();
+        }
     }
 }
