+++ migrations/Version20210802174320.php
@@ -26,17 +26,25 @@ final class Version20210802174320 extends AbstractMigration
 
     public function up(Schema $schema): void
     {
-        $teams = $schema->getTable('kimai2_teams');
-        // @see https://github.com/kimai/kimai/issues/2706
-        if ($teams->hasForeignKey('FK_3BEDDC7F8F7DE5D7')) {
-            $teams->removeForeignKey('FK_3BEDDC7F8F7DE5D7');
+        if ($this->isPlatformMySQL()) {
+            $teams = $schema->getTable('kimai2_teams');
+            // @see https://github.com/kimai/kimai/issues/2706
+            if ($teams->hasForeignKey('FK_3BEDDC7F8F7DE5D7')) {
+                $teams->removeForeignKey('FK_3BEDDC7F8F7DE5D7');
+            }
+            $teams->dropColumn('teamlead_id');
+        } else {
+            $this->preventEmptyMigrationWarning();
         }
-        $teams->dropColumn('teamlead_id');
     }
 
     public function down(Schema $schema): void
     {
-        $teams = $schema->getTable('kimai2_teams');
-        $teams->addColumn('teamlead_id', 'integer', ['length' => 11, 'notnull' => false]);
+        if ($this->isPlatformMySQL()) {
+            $teams = $schema->getTable('kimai2_teams');
+            $teams->addColumn('teamlead_id', 'integer', ['length' => 11, 'notnull' => false]);
+        } else {
+            $this->preventEmptyMigrationWarning();
+        }
     }
 }
